finished list of inventory items

+++ TeachersPet/src/teacherspet/Battling/CaptureItem.java
@@ -1,10 +1,12 @@
 public class CaptureItem extends Item{
     private double success;
 
-    CaptureItem(String name, int success) {
+    CaptureItem(String name, double success) {
         super(name);
         this.success = success;
     }
 
-    public double getSuccess(){return success;}
+    public double getSuccess(){
+        return success;
+    }
 }
